+++ JsonMergePatch.sln
@@ -29,9 +29,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "JsonMergePatch.SourceGenera
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConsoleAppLibrary", "sample\ConsoleAppLibrary\ConsoleAppLibrary.csproj", "{3592FF83-D9D6-473A-97CE-94FF0D7FE5F1}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AspNetCoreWebApi2", "sample\AspNetCoreWebApi2\AspNetCoreWebApi2.csproj", "{658A69D6-C5EF-490D-AC3E-AB2ADAFECE4A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AspNetCoreMinimal", "sample\AspNetCoreMinimal\AspNetCoreMinimal.csproj", "{882E1027-9F66-4779-B85C-BC10C38E6816}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AspNetCoreWebApi2.Entities", "sample\AspNetCoreWebApi2.Entities\AspNetCoreWebApi2.Entities.csproj", "{8A610CE4-0640-4674-AB66-42121026E093}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AspNetCoreMinimal.Entities", "sample\AspNetCoreMinimal.Entities\AspNetCoreMinimal.Entities.csproj", "{D9F5C75F-DD5F-4B0A-A7C2-BD928EDC4838}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -87,14 +87,14 @@ Global
 		{3592FF83-D9D6-473A-97CE-94FF0D7FE5F1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3592FF83-D9D6-473A-97CE-94FF0D7FE5F1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3592FF83-D9D6-473A-97CE-94FF0D7FE5F1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{658A69D6-C5EF-490D-AC3E-AB2ADAFECE4A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{658A69D6-C5EF-490D-AC3E-AB2ADAFECE4A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{658A69D6-C5EF-490D-AC3E-AB2ADAFECE4A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{658A69D6-C5EF-490D-AC3E-AB2ADAFECE4A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{8A610CE4-0640-4674-AB66-42121026E093}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{8A610CE4-0640-4674-AB66-42121026E093}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{8A610CE4-0640-4674-AB66-42121026E093}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{8A610CE4-0640-4674-AB66-42121026E093}.Release|Any CPU.Build.0 = Release|Any CPU
+		{882E1027-9F66-4779-B85C-BC10C38E6816}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{882E1027-9F66-4779-B85C-BC10C38E6816}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{882E1027-9F66-4779-B85C-BC10C38E6816}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{882E1027-9F66-4779-B85C-BC10C38E6816}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D9F5C75F-DD5F-4B0A-A7C2-BD928EDC4838}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D9F5C75F-DD5F-4B0A-A7C2-BD928EDC4838}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D9F5C75F-DD5F-4B0A-A7C2-BD928EDC4838}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D9F5C75F-DD5F-4B0A-A7C2-BD928EDC4838}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -103,8 +103,8 @@ Global
 		{910E8728-872B-43EF-994F-B67953874BCB} = {0E750D74-69C7-433C-92C7-3F026083C255}
 		{1FE80A1A-9402-4953-9A7F-F01B1D0539B5} = {0E750D74-69C7-433C-92C7-3F026083C255}
 		{3592FF83-D9D6-473A-97CE-94FF0D7FE5F1} = {0E750D74-69C7-433C-92C7-3F026083C255}
-		{658A69D6-C5EF-490D-AC3E-AB2ADAFECE4A} = {0E750D74-69C7-433C-92C7-3F026083C255}
-		{8A610CE4-0640-4674-AB66-42121026E093} = {0E750D74-69C7-433C-92C7-3F026083C255}
+		{882E1027-9F66-4779-B85C-BC10C38E6816} = {0E750D74-69C7-433C-92C7-3F026083C255}
+		{D9F5C75F-DD5F-4B0A-A7C2-BD928EDC4838} = {0E750D74-69C7-433C-92C7-3F026083C255}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {55007954-1D8A-4553-A15C-FE4AE17C0A4B}
