+++ sample/AspNetCoreMinimal.Entities/AspNetCoreMinimal.Entities.csproj
