+++ sample/AspNetCoreMinimal/appsettings.Development.json
