+++ sample/AspNetCoreMinimal/appsettings.json
